Replacing ghcjs-vdom with BeaufiulDestinations/virtual-dom
The API is essentially the same, but without the overloading magic of the old
library. I’ve refactored the current codebase to follow these conventions.

Overview of changes

- Imports:
  GHCJS.VDOM.Element -> Web.VirtualDom.Html
  GHCJS.VDOM.Attribute -> Web.VirtualDom.Html.Attributes
  GHCJS.VDOM.Event -> Web.VirtualDom.Html.Events

- All custom element/attributes (i.e. not defined in the above modules) should
be created with the functions Web.VirtualDom(node, attribute, property) and not the ugly custom… functions.

- It is longer required/possible to use () for attributes/children, or single
element in place of child list. The latter can give yous errors such as
couldn't match [Html] with Html and vice versa.

EXAMPLE

Old style:
  p () [text ”hello”]
  br () ()
  h1 () (text ”Header”)

New style:
  p [] [text ”hello”]
  br [] []
  h1 [] [text ”Header”]